Separate ProximateSeasons into component
