+++ src/components/ProximateSeasons/index.js
@@ -1,0 +1,3 @@
+import ProximateSeasons from './ProximateSeasons';
+
+export default ProximateSeasons;
